- Apparently, using unsigned int for the index instead of unsigned short breaks the output for DirectX. - Time lost = 7 days!

+++ Engine/Mesh.cpp
@@ -44,7 +44,7 @@ bool Mesh::AddVertex(Vertex v)
 
 /******************************************************************************************************************/
 
-bool Mesh::AddIndex(unsigned int i)
+bool Mesh::AddIndex(unsigned short i)
 {
 	if (!_locked)
 	{
